learning flecs with C api heh
