--- src/Engine/Components.h
@@ -1,19 +1,0 @@
-#pragma once
-
-
-namespace rosy {
-
-	// Components
-	struct c_position
-	{
-		float x{ 0.f };
-		float y{ 0.f };
-		float z{ 0.f };
-	};
-
-	// Tags
-
-	struct t_mob {};
-	struct t_rosy {};
-	struct t_floor {};
-}
