+++ src/Engine/Engine.vcxproj
@@ -256,7 +256,6 @@
     <ClInclude Include="..\libs\imgui\imstb_truetype.h" />
     <ClInclude Include="..\Packager\Asset.h" />
     <ClInclude Include="Camera.h" />
-    <ClInclude Include="Components.h" />
     <ClInclude Include="Engine.h" />
     <ClInclude Include="Level.h" />
     <ClInclude Include="Graphics.h" />
